Make ShareMode just use a History object directly, instead of defining its own Downloads class
